Fix array keys can be int or string, not float

+++ src/Benchmark/HttpTimeDataCollector.php
@@ -100,7 +100,7 @@ class HttpTimeDataCollector implements \IteratorAggregate
     {
         sort($arr);
 
-        return $arr[round($percentile * count($arr) - 1.0 - $percentile)];
+        return $arr[(int) round($percentile * count($arr) - 1.0 - $percentile)];
     }
 
     private function guardAtLeast2Results()
